requirements.txt
python-dotenv==0.15.0
beautifulsoup4==4.9.3